// frontend/src/api/model/chargeCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type ChargeCategory = typeof ChargeCategory[keyof typeof ChargeCategory];


export const ChargeCategory = {
  HOSPITALIZATION_DAY: 'HOSPITALIZATION_DAY',
  INTENSIVE_CARE_DAY: 'INTENSIVE_CARE_DAY',
  MEDICATION: 'MEDICATION',
  LABORATORY: 'LABORATORY',
  IMAGING: 'IMAGING',
  PROCEDURE: 'PROCEDURE',
  SUPPLY: 'SUPPLY',
  PROFESSIONAL_FEE: 'PROFESSIONAL_FEE',
  OTHER: 'OTHER',
} as const;