// frontend/src/api/model/nomenclador.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Coding system the practice belongs to.
 */
export type Nomenclador = typeof Nomenclador[keyof typeof Nomenclador];


export const Nomenclador = {
  NACIONAL: 'NACIONAL',
  NBU: 'NBU',
  NU_SSS: 'NU_SSS',
  HPGD: 'HPGD',
  PROPIO: 'PROPIO',
} as const;